refactor: Extract global error handler logic
